+++ libraries/Geometry/src/Intersections.cpp
@@ -6,7 +6,7 @@
 #include <memory>
 
 #include "Intersections.hpp"
-#include "FloatUtils.hpp"
+#include "floatUtils.hpp"
 #include "Shape.hpp"
 #include "Transformations.hpp"
 
@@ -67,7 +67,7 @@ Computations prepareComputations(Intersection intersection, const utility::Ray& 
  * @return The smallest positive intersection, wrapped in std::optional. If no positive intersection is found, returns std::nullopt.
  */
 std::optional<Intersection> hit(const std::vector<Intersection>& intersections) noexcept{
-  auto positiveT = intersections | std::views::filter([](const auto& intersection) { return intersection.dist > 0.0; });
+  auto positiveT = intersections | std::views::filter([](const auto& intersection) { return intersection.dist > 0.0f; });
   auto smallestIntersection = std::ranges::min_element(positiveT, {}, [](const auto& intersection){return intersection.dist;});
 
   if(smallestIntersection == positiveT.end())
@@ -75,7 +75,7 @@ std::optional<Intersection> hit(const std::vector<Intersection>& intersections) 
   return *smallestIntersection;
 }
 
-double schlick(const Computations& comps) noexcept{
+float schlick(const Computations& comps) noexcept{
   auto cos = comps.eyeVector.dot(comps.normalVector);
   if(comps.n1 > comps.n2){
     auto n = comps.n1 / comps.n2;
