+++ libraries/Utility/src/Tuple.cpp
@@ -5,16 +5,16 @@
 #include <execution>
 
 #include "Tuple.hpp"
-#include "FloatUtils.hpp"
+#include "floatUtils.hpp"
 
 namespace raytracer {
 namespace utility { 
 
 bool Tuple::isVector() const noexcept {
-  return floatNearlyEqual(this->w(), 0.0);
+  return floatNearlyEqual(this->w(), 0.0f);
 }
 bool Tuple::isPoint() const noexcept {
-  return floatNearlyEqual(this->w(), 1.0);
+  return floatNearlyEqual(this->w(), 1.0f);
 }
 
 bool Tuple::operator==(const Tuple& rhs) const noexcept {
@@ -31,7 +31,7 @@ Tuple Tuple::operator+(const Tuple& rhs) const noexcept {
   return Tuple(this->x()+rhs.x(), this->y()+rhs.y(), 
                this->z()+rhs.z(), this->w()+rhs.w());
 }
-Tuple Tuple::operator+(const double& rhs) const noexcept {
+Tuple Tuple::operator+(const float& rhs) const noexcept {
   return Tuple(this->x()+rhs, this->y()+rhs, 
                this->z()+rhs, this->w()+rhs);
 }
@@ -40,11 +40,11 @@ Tuple Tuple::operator-(const Tuple& rhs) const noexcept{
   return Tuple(this->x()-rhs.x(), this->y()-rhs.y(), 
                this->z()-rhs.z(), this->w()-rhs.w());
 }
-Tuple Tuple::operator*(const double& rhs) const noexcept{
+Tuple Tuple::operator*(const float& rhs) const noexcept{
   return Tuple(this->x()*rhs, this->y()*rhs, 
                this->z()*rhs, this->w()*rhs);
 }
-Tuple Tuple::operator/(const double& rhs) const noexcept{
+Tuple Tuple::operator/(const float& rhs) const noexcept{
   return Tuple(this->x()/rhs, this->y()/rhs, 
                this->z()/rhs, this->w()/rhs);
 }
@@ -52,9 +52,9 @@ Tuple Tuple::operator-() const noexcept{
   return *this*-1;
 }
 
-double Tuple::magnitude() const noexcept{
+float Tuple::magnitude() const noexcept{
   auto sumSquared = std::transform_reduce(std::execution::unseq, this->data.cbegin(), this->data.cend(), 
-                                          0.0, std::plus<>(), [](const auto& val){return val*val;});
+                                          0.0f, std::plus<>(), [](const auto& val){return val*val;});
   return std::sqrt(sumSquared);
 }
 
@@ -62,9 +62,9 @@ Tuple Tuple::normalize() const noexcept{
   return *this/this->magnitude();
 }
 
-double Tuple::dot(const Tuple& rhs) const noexcept{
+float Tuple::dot(const Tuple& rhs) const noexcept{
   return std::transform_reduce(std::execution::unseq, this->data.cbegin(), this->data.cend(), 
-                               rhs.data.cbegin(), 0.0, std::plus<>(), std::multiplies<>());
+                               rhs.data.cbegin(), 0.0f, std::plus<>(), std::multiplies<>());
 }
 
 Tuple Tuple::cross(const Tuple& rhs) const noexcept{
@@ -77,10 +77,10 @@ Tuple Tuple::reflect(const Tuple& normal) const noexcept{
   return *this - (normal * 2 * this->dot(normal));
 }
 
-Tuple Point(const double x, const double y, const double z) noexcept{
+Tuple Point(const float x, const float y, const float z) noexcept{
   return Tuple(x,y,z,1);
 }
-Tuple Vector(const double x, const double y, const double z) noexcept{
+Tuple Vector(const float x, const float y, const float z) noexcept{
   return Tuple(x,y,z,0);
 }
